refactor: expand Rector to Streaming / Public\Controllers / Ministra (stage 2)
Widen the Rector config to the remaining class-based trees (Streaming,
Public\Controllers, Ministra), keeping the templates, procedural
front-controllers (Public/stream, Public/admin, Ministra/portal.php) and the
streaming hot-path bootstraps out. Also document the recurring dropped-parens
bug + the review grep in the config header.

The resulting 107-file pass (94 Public\Controllers, 11 Streaming, 2 Ministra)
was reviewed against the suite + PHPStan + the usual scans:
- No dropped-parens bug this time (the assignment-in-condition pattern didn't
  occur in these files).
- 6 locally-called private static helpers became instance methods
  (behaviour-preserving; all called via $this->, no static call sites); one
  unused private param dropped (FanoutConfig::desired $rSnapshot, call site
  updated).
- De Morgan / ternary / dead-code simplifications; two `$x = getById()`
  truthy-assignments folded into the condition (no comparison, safe).

PHPStan level 5 clean; suite 721 tests / 0 errors.

+++ src/Streaming/Fanout/FanoutClient.php
@@ -606,7 +606,7 @@ class FanoutClient {
 			curl_setopt($rCurl, CURLOPT_HTTPHEADER, ['Content-Type: application/json']);
 		}
 		$rResponse = curl_exec($rCurl);
-		$rCode = (int) curl_getinfo($rCurl, CURLINFO_HTTP_CODE);
+		$rCode = curl_getinfo($rCurl, CURLINFO_HTTP_CODE);
 		$rErrno = curl_errno($rCurl);
 		curl_close($rCurl);
 		return ['code' => $rCode, 'body' => is_string($rResponse) ? $rResponse : null, 'errno' => $rErrno];
